requirements3.txt
# =============================================================================
# REQUIREMENTS FOR MAIN2.PY - Complete Integration System
# =============================================================================
# This file includes all dependencies for:
# - main2.py (main integration script)
# - core/vision.py (YOLO + MiDaS depth estimation)
# - utils/ocr_engine.py (PaddleOCR text detection)
# - utils/audio_input_deep.py (Deepgram speech-to-text)
# - utils/audio_output.py (ElevenLabs TTS with spatial audio)
# - utils/interpret.py (Google Gemini LLM)
# =============================================================================

# Core Computer Vision (from cv_requirements.txt)
# -----------------------------------------------------------------------------
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.0
polars==1.38.1
polars-runtime-32==1.38.1
psutil==7.2.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
safetensors==0.7.0
scipy==1.17.0
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
timm==1.0.24
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
typer-slim==0.21.1
typing_extensions==4.15.0
ultralytics==8.4.12
ultralytics-thop==2.0.18
urllib3==2.6.3

# OCR Engine (PaddleOCR)
# -----------------------------------------------------------------------------
paddleocr>=2.8.0
# Note: paddlepaddle installation is platform-specific
# For CPU: pip install paddlepaddle
# For GPU: pip install paddlepaddle-gpu
# It will be installed automatically as a dependency of paddleocr

# Audio Input (Deepgram)
# -----------------------------------------------------------------------------
sounddevice>=0.5.0
deepgram-sdk==2.12.0  # Using v2 for compatibility with existing code

# Audio Output (ElevenLabs + pydub)
# -----------------------------------------------------------------------------
elevenlabs>=2.30.0
pydub>=0.25.1

# Google Gemini LLM
# -----------------------------------------------------------------------------
google-generativeai>=0.8.0

# Environment Variables
# -----------------------------------------------------------------------------
python-dotenv>=1.0.0

# Additional Dependencies
# -----------------------------------------------------------------------------
# These may be implicitly required by the above packages or for full functionality
protobuf==5.29.3
